+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Custom GTM dataLayer event for when user submits a review
+
 ## [3.11.2] - 2022-10-20
 
 ### Changed